+++ src/components/textarea.js
@@ -36,8 +36,30 @@ module.exports = function(app) {
             }
             return 'formio/components/texteditor.html';
           }
+          if ($scope.readOnly && $scope.component.wysiwyg) {
+            return 'formio/componentsView/content.html';
+          }
           return 'formio/components/textarea.html';
         },
+        viewTemplate: function($scope) {
+          if ($scope.readOnly && $scope.component.wysiwyg) {
+            return 'formio/componentsView/content.html';
+          }
+          else {
+            return 'formio/components/textarea.html';
+          }
+        },
+        controller: [
+          '$scope',
+          '$sanitize',
+          function($scope, $sanitize) {
+            if ($scope.readOnly && $scope.component.wysiwyg) {
+              $scope.$watch('data.' + $scope.component.key, function() {
+                $scope.html = $sanitize($scope.data[$scope.component.key]);
+              });
+            }
+          }
+        ],
         settings: {
           input: true,
           tableView: true,
@@ -72,6 +94,9 @@ module.exports = function(app) {
               FormioUtils) {
       $templateCache.put('formio/components/textarea.html', FormioUtils.fieldWrap(
         fs.readFileSync(__dirname + '/../templates/components/textarea.html', 'utf8')
+      ));
+      $templateCache.put('formio/componentsView/content.html', FormioUtils.fieldWrap(
+        fs.readFileSync(__dirname + '/../templates/componentsView/content.html', 'utf8')
       ));
       $templateCache.put('formio/components/texteditor.html', FormioUtils.fieldWrap(
         fs.readFileSync(__dirname + '/../templates/components/texteditor.html', 'utf8')